+++ DataImporter/test/unit/src/de/unibielefeld/gi/kotte/laborprogramm/dataImporter/GelDataReaderTest.java
@@ -5,6 +5,7 @@ import de.unibielefeld.gi.kotte.laborprogramm.proteomik.api.gel.IGelFactory;
 import de.unibielefeld.gi.kotte.laborprogramm.xml.gelData.GelData;
 import de.unibielefeld.gi.kotte.laborprogramm.xml.gelData.GelImage;
 import java.io.File;
+import javax.swing.JFileChooser;
 import org.openide.util.Lookup;
 
 /**
@@ -13,11 +14,22 @@ import org.openide.util.Lookup;
  */
 public class GelDataReaderTest {
 
+    private final static boolean FILECHOOSER = true;
+
     public static void main(String[] args) {
-        File f = new File("gelImages.xml");
-        f.deleteOnExit();
-        String path = "/de/unibielefeld/gi/kotte/laborprogramm/dataImporter/resources/gelImages.xml";
-        ResourceHandler.writeResourceToDisk(path, f);
+        File f = null;
+        if (!FILECHOOSER) {
+            f = new File("Export_1.xlsx");
+            f.deleteOnExit();
+            String path = "/de/unibielefeld/gi/kotte/laborprogramm/dataImporter/resources/gelImages.xml";
+            ResourceHandler.writeResourceToDisk(path, f);
+        } else {
+            JFileChooser jfc = new JFileChooser();
+            int status = jfc.showOpenDialog(null);
+            if (status == JFileChooser.APPROVE_OPTION) {
+                f = jfc.getSelectedFile();
+            }
+        }
         GelDataReader gdr = new GelDataReader();
         GelData gd = gdr.getGelData(f);
         for (GelImage gi : gd.getGelImages().getGelImage()) {
